AFTER FACTURATION CREATE

+++ src/app/views/ahmedbaba/classes/parametrages/parametragelist/parametragelist.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
+import { ConfirmationdialogComponent } from 'app/views/ahmedbaba/confirmationdialog/confirmationdialog.component';
 import { ClasseService } from 'app/views/ahmedbaba/services/classe.service';
 import { ParametrageService } from 'app/views/ahmedbaba/services/parametrage.service';
 
@@ -16,7 +18,8 @@ listParametrages: any[] = [];
     private classeService: ClasseService,
     private parametrageService: ParametrageService,
     private router: Router,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+        private dialog: MatDialog
   ) { }
 
   ngOnInit(): void {
@@ -29,8 +32,7 @@ listParametrages: any[] = [];
     this.classeService.getParametragesClasse(idClasse).subscribe({
       next: (data: any) => {
         console.log("Data : ", data)
-        this.listParametrages = data;
-        console.log("PARAMETRAGES FACTURE :", data);
+        this.listParametrages = data; 
       },
       error: (err) => {
         console.error("Erreur récupération paramétrages :", err);
@@ -42,15 +44,24 @@ listParametrages: any[] = [];
     this.router.navigate(['/ahmedbaba/classes/parametrages/update', id]);
   }
 
-  delete(id: number): void {
-    // Appel service suppression
-    this.parametrageService.deleteParametrage(id).subscribe({
-      next: () => {
-        this.listParametrages = this.listParametrages.filter(p => p.id !== id);
-        console.log("Paramétrage supprimé :", id);
-      },
-      error: (err) => console.error("Erreur suppression :", err)
-    });
-  }
+delete(id: number): void {
 
+  const dialog = this.dialog.open(ConfirmationdialogComponent, {
+    width: '350px',
+    data: { message: "Voulez-vous vraiment supprimer ce paramétrage ?" }
+  });
+
+  dialog.afterClosed().subscribe(result => {
+    if (result === true) {
+      this.parametrageService.deleteParametrage(id).subscribe({
+        next: () => {
+          this.listParametrages = this.listParametrages.filter(p => p.id !== id);
+          console.log("Paramétrage supprimé :", id);
+        },
+        error: (err) => console.error("Erreur suppression :", err)
+      });
+    }
+  });
+
+}
 }
